refactor: 폴더 구조 리팩토링(#112)
- features/exams/components -> shared로 이동
- features/exams/components 폴더 삭제
- features/exams/components/story -> shared로 story 파일 이동
- story 폴더 삭제
- 배럴 패턴 적용
