+++ src/features/exams/index.ts
@@ -1,15 +1,15 @@
 export { default as EmptyState } from './shared/components/EmptyState.tsx'
-export { default as ExamList } from './components/ExamList.tsx'
-export { ExamColumns } from './components/examConfig.tsx'
+export { default as ExamList } from './shared/components/ExamList.tsx'
+export { ExamColumns } from './shared/components/examConfig.tsx'
 export { mockExamList } from '../../mocks/mockExamList.ts'
-export { default as ExamCreateModal } from './components/ExamCreateModal.tsx'
+export { default as ExamCreateModal } from './shared/components/ExamCreateModal.tsx'
 export { createExamDropdown } from './utils/createExamDropdown.ts'
 export type { ExamQuestion, ExamQuestionResponse } from './types.ts'
 export { ExamQuestionDetailBody } from './shared/components/detail-modal/ExamQuestionDetailBody.tsx'
 export { ExamQuestionDetailFooter } from './shared/components/detail-modal/ExamQuestionDetailFooter.tsx'
 export { ExamQuestionDetailSide } from './shared/components/detail-modal/ExamQuestionDetailSide.tsx'
 export { default as ExamQuestionDetailModal } from './shared/components/detail-modal/ExamQuestionDetailModal.tsx'
-export { default as ExamDeploymentsModal } from './components/ExamDeploymentsModal.tsx'
+export { default as ExamDeploymentsModal } from './shared/components/ExamDeploymentsModal.tsx'
 export { default as ExamPage } from '../../pages/ExamManagementPage.tsx'
 export * from './exam-managements/questions/types.ts'
 export { default as ExamDeletePopupModal } from './exam-managements/questions/components/ExamDeletePopupModal.tsx'
@@ -22,4 +22,6 @@ export { useExamDeleteMutation } from './queries/useExamDeleteMutation.ts'
 export { useExamDeploymentsMutation } from './queries/useExamDeploymentsMutation.ts'
 
 export { examDeploymentsSchema } from './schemas/examDeploymentsSchema.ts'
-export { default as DeploymentHistoryModal } from './components/DeploymentHistoryModal.tsx'
+export { default as DeploymentHistoryModal } from './shared/components/DeploymentHistoryModal.tsx'
+export * from './shared/components/deploymentHistoryModalConfig.tsx'
+export * from './types.ts'
