+++ src/features/exams/shared/components/detail-modal/ExamQuestionDetail.stories.tsx
