+++ src/features/exams/submissions/components/SubmissionList.tsx
@@ -1,8 +1,8 @@
+import type { Submission } from '@features/exams'
+
 import { DataTableLayout, type DataTableLayoutProps } from '@components'
 import { PAGE_SIZE } from '@constants'
 import { getCoreRowModel, useReactTable } from '@tanstack/react-table'
-
-import type { Submission } from '../../types.ts'
 
 import { SubmissionColumns } from './submissionConfig.tsx'
