+++ src/mocks/examDropdown.ts
@@ -1,4 +1,4 @@
-import { createExamDropdown } from '../features/exams/utils/createExamDropdown'
+import { createExamDropdown } from '@features/exams'
 
 /**
  * value : 리스트 값
